Fix align function
Align function was allowing the starting alignment state to be set to undefined, which is later used to attempt to access an array index, causing crashes. Intention appears to allow for undefined values, so some sort of handling should likely be done. From what I can intuit, this appears as if it would be the preferred way to handle this sort of input to this function